# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivityEntry": {
        "description": "A single activity log entry.",
        "properties": {
          "action": {
            "description": "Short action description",
            "title": "Action",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional structured data",
            "title": "Details"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Duration of the action in ms",
            "title": "Duration Ms"
          },
          "page": {
            "default": "",
            "description": "Page where the action occurred",
            "title": "Page",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "ActivityEntry",
        "type": "object"
      },
...
      "DrugInfo": {
        "properties": {
          "affinity_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affinity Type"
          },
          "affinity_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affinity Value"
          },
          "chembl_id": {
            "title": "Chembl Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "smiles": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smiles"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "name",
          "chembl_id",
          "source"
        ],
        "title": "DrugInfo",
        "type": "object"
      },
...
      "FetchPDBRequest": {
        "properties": {
          "output_dir": {
            "title": "Output Dir",
            "type": "string"
          },
          "pdb_id": {
            "description": "4-character PDB accession code",
            "maxLength": 4,
            "minLength": 4,
            "title": "Pdb Id",
            "type": "string"
          }
        },
        "required": [
          "pdb_id",
          "output_dir"
        ],
        "title": "FetchPDBRequest",
        "type": "object"
      },
      "FetchPDBResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "pdb_path": {
            "title": "Pdb Path",
            "type": "string"
          }
        },
        "required": [
          "pdb_path",
          "message"
        ],
        "title": "FetchPDBResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UniprotInfo": {
        "properties": {
          "accession": {
            "title": "Accession",
            "type": "string"
          },
          "bio_processes": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Bio Processes",
            "type": "array"
          },
          "chembl_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chembl Id"
          },
          "entry_name": {
            "default": "Unknown",
            "title": "Entry Name",
            "type": "string"
          },
          "function": {
            "title": "Function",
            "type": "string"
          },
          "gene": {
            "title": "Gene",
            "type": "string"
          },
          "network_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network Image Url"
          },
          "organism": {
            "title": "Organism",
            "type": "string"
          },
          "pathways": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Pathways",
            "type": "array"
          }
        },
        "required": [
          "accession",
          "gene",
          "organism",
          "function"
        ],
        "title": "UniprotInfo",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Drug Discovery Suite — REST + WebSocket API",
    "title": "Helix Core Backend",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/activity/list": {
      "get": {
        "description": "Read activity entries with pagination and optional page filter.",
        "operationId": "list_activities_api_activity_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "filterPage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filterpage"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Activities",
        "tags": [
          "Activity Log"
        ]
      }
    },
    "/api/activity/log": {
      "post": {
        "description": "Append an activity entry to the log file.",
        "operationId": "log_activity_api_activity_log_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Activity",
        "tags": [
          "Activity Log"
        ]
      }
    },
...
    "/api/fetch/controls/{pdb_id}": {
      "get": {
        "description": "Search ChEMBL for approved drugs or high-affinity binders.",
        "operationId": "find_drug_controls_api_fetch_controls__pdb_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pdb_id",
            "required": true,
            "schema": {
              "title": "Pdb Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DrugInfo"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Find Drug Controls Api Fetch Controls  Pdb Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find Drug Controls",
        "tags": [
          "PDB Fetcher"
        ]
      }
    },
    "/api/fetch/pdb": {
      "post": {
        "description": "Download a PDB from RCSB and save the unmodified structure.",
        "operationId": "fetch_pdb_api_fetch_pdb_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchPDBRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchPDBResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch Pdb",
        "tags": [
          "PDB Fetcher"
        ]
      }
    },
    "/api/fetch/uniprot/{pdb_id}": {
      "get": {
        "description": "Fetch UniProt data for a PDB accession.",
        "operationId": "get_uniprot_info_api_fetch_uniprot__pdb_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pdb_id",
            "required": true,
            "schema": {
              "title": "Pdb Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniprotInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Uniprot Info",
        "tags": [
          "PDB Fetcher"
        ]
      }
    },
...
  }
}